Move PlayerInput,InputKey, Settings to own package
These files do not fit within the controller package, and the
eventhandler package is greatly redundant.

< organisation
