+++ src/main/java/nl/tudelft/controller/StartScreenStateController.java
@@ -8,8 +8,9 @@ import nl.tudelft.model.Game;
 import nl.tudelft.model.MultiplayerGame;
 import nl.tudelft.model.Player;
 import nl.tudelft.model.SingleplayerGame;
-
+import nl.tudelft.settings.Settings;
 import nl.tudelft.view.StartScreenState;
+
 import org.newdawn.slick.Color;
 import org.newdawn.slick.GameContainer;
 import org.newdawn.slick.Graphics;
