+++ src/main/java/nl/tudelft/settings/PlayerInput.java
@@ -1,4 +1,4 @@
-package nl.tudelft.controller.eventhandlers;
+package nl.tudelft.settings;
 
 import java.util.Observable;
 import java.util.Observer;
